+++ Sources/Options.swift
@@ -97,7 +97,7 @@ public struct TransitionOptions {
     /// Regular view animation options. Default value is `curveEaseInOut`.
     public var animationOptions: UIViewAnimationOptions = .curveEaseInOut
 
-    public init(duration: TimeInterval = 0.5, contentScale: CGFloat = 0.88, visibleContentWidth: CGFloat = 56.0) {
+    public init(duration: TimeInterval = 0.5, contentScale: CGFloat = 0.86, visibleContentWidth: CGFloat = 56.0) {
         self.duration = duration
         self.contentScale = contentScale
         self.visibleContentWidth = visibleContentWidth